server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.10,<4.0"
dependencies = [
    "celery (==5.4.0)",
    "django (==5.1.6)",
    "djangorestframework (==3.15.2)",
    "drf-yasg (==1.21.8)",
    "django-cors-headers (==4.7.0)",
    "pendulum (>=3.0.0,<4.0.0)",
    "typer (>=0.15.2,<0.16.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "django-celery-beat (>=2.7.0,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "netcdf4 (>=1.7.2,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "django-import-export (>=4.3.7,<5.0.0)",
    "django-celery-results (>=2.5.1,<3.0.0)",
    "pygrid (>=0.0.1,<0.0.2)",
    "pygrib (>=2.1.6,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "geopandas (>=1.0.1,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "xarray (==2025.6.1)",
    "ninja (==1.11.1.4)"
]

[[tool.poetry.source]]
name = "custom"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
typeCheckingMode = "off"